Repace delegated with Z for linux.
The docker volume bind-mount option 'delegated' is specific
to macOS. The documentation says it's ignored on other platforms,
but on my machine (Fedora 30 x86_64) it throws an error.

I've removed it, but making the script portable with some kind
of os detection would work too.

Similarly, selinux requires the Z attribute on my system
to allow the bind mount into the container.